Сделал интерфейс-репозиторий для основной сущности и реализацию работающую с БД

+++ SimpleServiceJDBC/src/main/java/model/Course.java
@@ -9,6 +9,6 @@ public class Course {
 
     private String name;
 
-    private Byte duration;
+    private Integer duration;
 
 }
